refactor: remove confirm_close module, distribute logic into terminal and layout
Move close-confirmation helpers and dialog functions into their domain modules:
- terminal.rs: should_confirm_close, has_live_terminal, pane_has_live_terminal, dialog helpers
- layout/tab.rs: PendingTabClose, CloseConfirmed, process_pending_tab_closes
- layout/pane.rs: PendingPaneClose, process_pending_pane_closes
- os_menu.rs: PendingWindowClose, process_pending_window_close

Delete confirm_close.rs and ConfirmClosePlugin.
